#1092 [REFACTOR] footerData로 파일 분리
