+++ src/shared/constant/index.js
@@ -1,4 +1,5 @@
 export * from './board.js';
+export * from './footerData.js';
 export * from './icon.js';
 export * from './loading.js';
 export * from './majors.js';
